pokegen/requirements.txt
datasets 
huggingface_hub 
Pillow
wandb
accelerate>=0.22.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
datasets
peft==0.15.0